requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/requirements.in
#
amqp==5.0.1               # via kombu
asgiref==3.2.10           # via django
attrs==20.2.0
billiard==3.6.3.0         # via celery
celery-progress==0.0.14
celery[redis]==5.0.0
certifi==2020.11.8        # via requests
cffi==1.14.4              # via cryptography
chardet==3.0.4            # via requests
click-didyoumean==0.0.3   # via celery
click-repl==0.1.6         # via celery
click==7.1.2              # via celery, click-didyoumean, click-repl
coreapi==2.3.3
coreschema==0.0.4         # via coreapi
cryptography==3.2.1       # via pyjwt
defusedxml==0.6.0         # via python3-openid
dj-stripe==2.4.1
django-allauth==0.44.0
django-anymail[mailgun]==8.1
django==3.1.6
djangorestframework==3.12.1
idna==2.10                # via requests
importlib-metadata==2.0.0  # via kombu
itypes==1.2.0             # via coreapi
jinja2==2.11.2            # via coreschema
jsonfield==3.1.0          # via dj-stripe
kombu==5.0.2              # via celery
markupsafe==1.1.1         # via jinja2
oauthlib==3.1.0           # via requests-oauthlib
prompt-toolkit==3.0.8     # via click-repl
psycopg2-binary==2.8.6
pycparser==2.20           # via cffi
pyjwt[crypto]==1.7.1      # via django-allauth
python3-openid==3.2.0     # via django-allauth
pytz==2020.1              # via celery, django
redis==3.5.3              # via celery
requests-oauthlib==1.3.0  # via django-allauth
requests==2.24.0          # via coreapi, django-allauth, django-anymail, requests-oauthlib, stripe
six==1.15.0               # via click-repl, cryptography
sqlparse==0.4.1           # via django
stripe==2.54.0
uritemplate==3.0.1        # via coreapi
urllib3==1.25.11          # via requests
vine==5.0.0               # via amqp, celery
wcwidth==0.2.5            # via prompt-toolkit
zipp==3.3.0               # via importlib-metadata
